beta 7

+++ app/Http/Controllers/SmallBoxController.php
@@ -6,6 +6,7 @@ use App\SmallBox;
 use Illuminate\Http\Request;
 use App\Http\Requests\SmallBoxRequest;
 use App\Http\Resources\SmallBox\SmallBoxResource;
+use App\Http\Resources\SmallBox\SmallBoxDetailResource;
 use App\Http\Resources\SmallBox\SmallBoxCollection;
 use Illuminate\Support\Facades\DB;
 
@@ -42,24 +43,87 @@ class SmallBoxController extends ApiController
     	return new SmallBoxCollection($smallBoxes); 
     }
 
+    public function getSmallBoxesActives() 
+    {
+        $data = DB::table('small_boxes AS s')
+          ->join('users AS u', 's.user_id', '=', 'u.id')
+          ->leftjoin('amounts AS a', 's.id', '=', 'a.small_box_id')
+          ->select('s.id', 's.date_init', 's.start_amount', DB::raw('COALESCE(SUM(a.amount), 0) add_amount'), 'u.name')
+          ->where('s.state', 1)
+          ->groupBy('s.id')
+          ->get();
+
+        return $this->respond($data);
+    }
+
+    public function active($user) 
+    {
+        $query = $this->smallBox->getState($user);
+        if ($query) {
+            if ($query->state == 1) {
+                return $this->respond([
+                    'flag' => false,
+                    'account' => $query->account_id
+                ]);
+            }
+        }
+
+        return $this->respond([
+            'flag' => true,
+            'message' => message('MSG014'),
+        ]);
+    }
+
     public function show($id)
     {
         $smallBox = $this->smallBox->findOrFail($id);
-        return new SmallBoxResource($smallBox); 
+        return new SmallBoxResource($smallBox);
+    }
+
+    public function detail($id)
+    {
+        $smallBox = $this->smallBox->findOrFail($id);
+        return new SmallBoxDetailResource($smallBox); 
+    }
+
+    public function getExpenseSmallBox(Request $request)
+    {
+        $params  = json_decode($request->data, true);
+
+        $date_end = date('Y-m-d H:i:s');
+        $smallBox = $this->smallBox->where('id', $params['id'])->first();
+        if ($smallBox->state == 0) {
+            $date_end = $smallBox->date_end;
+        }
+
+        $data = DB::table('expenses AS e')
+          ->join('projects AS p', 'e.project_id', '=', 'p.id')
+          ->select('e.title', 'e.payment', 'e.date', 'e.amount', 'p.name as project')
+          ->where(function($query)  use ($params, $date_end) {
+            $query->where('e.account_id', $params['account'])
+                  ->where('e.user_id', $params['user'])
+                  ->where('e.created_at', '>=', $params['date_init'])
+                  ->where('e.created_at', '<=', $date_end)
+                  ->whereNull('e.deleted_at');
+            })
+          ->get();
+
+        return $this->respond($data);
     }
 
     public function store(SmallBoxRequest $request)
     {
         DB::beginTransaction();
         try {
+            $query = $this->smallBox->getState($request->input('user_id'));
+            if ($query && $query->state == 1) {
+                return response()->json([
+                    'message' => message('MSG013'),
+                ], 422);
+            }
+
             $smallBox = $this->smallBox->create($request->all());
-            // if (!empty($request->accounts)) {
-            //     $account = array();
-            //     foreach ($request->accounts as $key => $value) {
-            //         $account[$value['id']] = ['income' => $value['incomes'], 'expense' => $value['expenses'], 'cash' => $value['cash']];
-            //     }
-            //     $smallBox->accounts()->attach($account);
-            // } 
+
             DB::commit();
         } catch (\Exception $e) {
             DB::rollback();
@@ -73,14 +137,15 @@ class SmallBoxController extends ApiController
         DB::beginTransaction();
         try {
             $smallBox = $this->smallBox->find($id);
-            $smallBox->update($request->smallBox);
-            if (!empty($request->accounts)) {
-                $account = array();
-                foreach ($request->accounts as $key => $value) {
-                    $account[$value['id']] = ['income' => $value['incomes'], 'expense' => $value['expenses'], 'cash' => $value['cash']];
-                }
-                $smallBox->accounts()->sync($account);
-            } 
+
+            if ($request->input('state') == 0){
+                return response()->json([
+                    'message' => message('MSG015'),
+                ], 422);
+            }
+
+            $smallBox->update($request->all());
+
             DB::commit();
         } catch (\Exception $e) {
             DB::rollback();
@@ -93,6 +158,13 @@ class SmallBoxController extends ApiController
     {
         try {
             $smallBox = $this->smallBox::find($id);
+
+            if (!is_null($smallBox->date_end)){
+                return response()->json([
+                    'msg' => message('MSG016')
+                ]);
+            }
+
             $smallBox->delete();
         } catch (\Exception $e) {
             return $this->respondInternalError();
